refactor: i18n rename ru-RU to ru
