+++ src/shared/locales/app.js
@@ -6,7 +6,7 @@ import appLocaleFr from '@shared/locales/fr'
 import appLocaleJa from '@shared/locales/ja'
 import appLocaleKo from '@shared/locales/ko'
 import appLocalePtBr from '@shared/locales/pt-BR'
-import appLocaleRu from '@shared/locales/ru-RU'
+import appLocaleRu from '@shared/locales/ru'
 import appLocaleTr from '@shared/locales/tr'
 import appLocaleZhCN from '@shared/locales/zh-CN'
 import appLocaleZhTW from '@shared/locales/zh-TW'
