+++ src/shared/locales/ru/about.js
