+++ src/shared/locales/ru/app.js
