+++ src/shared/locales/ru/edit.js
