+++ src/shared/locales/ru/help.js
