+++ src/shared/locales/ru/index.js
