+++ src/shared/locales/ru/menu.js
