+++ src/shared/locales/ru/preferences.js
