+++ src/shared/locales/ru/subnav.js
