+++ src/shared/locales/ru/window.js
